+++ packages/malloy/src/lang/parse-malloy.ts
@@ -81,7 +81,7 @@ import {locationContainsPosition} from './utils';
 import {Tag} from '../tags';
 import {MalloyParseInfo} from './malloy-parse-info';
 import {walkForModelAnnotation} from './parse-tree-walkers/model-annotation-walker';
-import {walkForTablePath} from './parse-tree-walkers/find-table-path';
+import {walkForTablePath} from './parse-tree-walkers/find-table-path-walker';
 
 export type StepResponses =
   | DataRequestResponse
