+++ packages/malloy/src/lang/parse-tree-walkers/find-table-path-walker.ts
