+++ packages/malloy/src/malloy.ts
@@ -80,7 +80,7 @@ import {
 import {DateTime} from 'luxon';
 import {Tag, TagParse, TagParseSpec, Taggable} from './tags';
 import {Dialect, getDialect} from './dialect';
-import {PathInfo} from './lang/parse-tree-walkers/find-table-path';
+import {PathInfo} from './lang/parse-tree-walkers/find-table-path-walker';
 
 export interface Loggable {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
